Lần này test ver1

+++ src/red/s1/boss/Boss.java
@@ -494,7 +494,7 @@ public class Boss extends Player implements IBossNew, IBossOutfit {
     // end loop
 
     @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
         this.checkAnThan(plAtt);
         if (!this.isDie()) {
